Fixed protein run for new file handling; added info for constants.py; moved more file handling to proper place; cleaned up TODOs; updated badges (RIP 100 downloads)
