+++ QUBEKit/dihedrals.py
@@ -4,7 +4,7 @@ from QUBEKit.engines import Gaussian, OpenMM, PSI4, RDKit
 from QUBEKit.utils import constants
 from QUBEKit.utils.decorators import for_all_methods, timer_logger
 from QUBEKit.utils.exceptions import TorsionDriveFailed
-from QUBEKit.utils.helpers import make_and_change_into
+from QUBEKit.utils.file_handling import make_and_change_into
 
 from collections import OrderedDict
 from copy import deepcopy
