+++ QUBEKit/run.py
@@ -18,8 +18,8 @@ from QUBEKit.utils.constants import COLOURS
 from QUBEKit.utils.decorators import exception_logger
 from QUBEKit.utils.display import display_molecule_objects, pretty_print, pretty_progress
 from QUBEKit.utils.exceptions import HessianCalculationFailed, OptimisationFailed
-from QUBEKit.utils.helpers import append_to_log, generate_bulk_csv, make_and_change_into, mol_data_from_csv, unpickle, \
-    update_ligand
+from QUBEKit.utils.file_handling import make_and_change_into
+from QUBEKit.utils.helpers import append_to_log, generate_bulk_csv, mol_data_from_csv, unpickle, update_ligand
 
 import argparse
 from collections import OrderedDict
