+++ QUBEKit/utils/constants.py
@@ -9,34 +9,34 @@ Division will give the wrong conversion.
 from collections import namedtuple
 
 
-AVOGADRO = 6.02214179e23        # Particles in 1 Mole
-ROOM_TEMP = 298.15              # Kelvin
-ROOM_PRESSURE = 101325          # Pascals
+AVOGADRO = 6.02214179e23                    # Particles in 1 Mole
+ROOM_TEMP = 298.15                          # Kelvin
+ROOM_PRESSURE = 101325                      # Pascals
 VACUUM_PERMITTIVITY = 8.8541878128e-12      # Farads per Metre
 ELECTRON_CHARGE = 1.602176634e-19           # Coulombs
-KB_KCAL_P_MOL_K = 0.0019872041      # Boltzmann constant in KCal/(mol * K)
+KB_KCAL_P_MOL_K = 0.0019872041              # Boltzmann constant in KCal/(mol * K)
 
-PI = 3.141592653589793
-DEG_TO_RAD = PI / 180
-RAD_TO_DEG = 180 / PI
+PI = 3.141592653589793                      # Pi
+DEG_TO_RAD = PI / 180                       # Degrees to radians
+RAD_TO_DEG = 180 / PI                       # Radians to degrees
 
 
-KCAL_TO_KJ = 4.184
-KJ_TO_KCAL = 0.23900573613
-J_TO_KCAL = 0.0002390057
+KCAL_TO_KJ = 4.184                          # Kilocalories to kiloJoules
+KJ_TO_KCAL = 0.23900573613                  # KiloJoules to kilocalories
+J_TO_KCAL = 0.0002390057                    # Joules to kilocalories
 
-J_TO_KCAL_P_MOL = J_TO_KCAL * AVOGADRO
+J_TO_KCAL_P_MOL = J_TO_KCAL * AVOGADRO      # Joules to kilocalories per mole
 
-HA_TO_KCAL_P_MOL = 627.509391
-KCAL_P_MOL_TO_HA = 0.00159360164
+HA_TO_KCAL_P_MOL = 627.509391               # Hartrees to kilocalories per mole
+KCAL_P_MOL_TO_HA = 0.00159360164            # Kilocalories per mole to Hartrees
 
-NM_TO_ANGS = 10
-ANGS_TO_NM = 0.1
+NM_TO_ANGS = 10                             # Nanometres to Angstroms
+ANGS_TO_NM = 0.1                            # Angstroms to nanometres
 
-ANGS_TO_M = 10e-10
+ANGS_TO_M = 10e-10                          # Angstroms to metres
 
-BOHR_TO_ANGS = 0.529177
-ANGS_TO_BOHR = 1.88972687777
+BOHR_TO_ANGS = 0.529177                     # Bohrs to Angstroms
+ANGS_TO_BOHR = 1.88972687777                # Angstroms to Bohrs
 
 
 # Used for printing colours to terminal. Wrap a colour and end around a block like so:
